feat(afternote): 삭제 액션 바 Variant5 dual-action 적용·SongPlaylistScreen 파일 분리
- core:ui AfternoteButton에 onSecondaryClick 추가 — Variant5+secondaryText 조합에서
  좌/우 절반이 독립 클릭 타깃인 dual-action 모드, 미지정 시 기존 단일 onClick 그대로
- MemorialPlaylistActionBar 커스텀 흰 배경 Row를 AfternoteButton(Variant5)로 교체
  (시안 정합: gray9 바·흰 라벨·중앙 divider)
- SearchableSongList에 동거하던 SongPlaylistScreen 3개 오버로드·설정 클래스를 전용 파일로 이동
- 스크린샷: 버튼 dual-action·플레이리스트 선택 모드 baseline 신규,
  기존 baseline은 docker(amd64) 사이클 재생성 (validate 통과)
